website-from-talents/requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-login
werkzeug
markdown2
